Updating function work mostly completed. working on arcades-edit view.

+++ src/main/java/com/gost/gaminggroup/gaminggroupapp/repository/ArcadeRepository.java
@@ -4,6 +4,6 @@ import com.gost.gaminggroup.gaminggroupapp.models.Arcade;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface ArcadeRepository extends JpaRepository<Arcade, Integer> {
+public interface ArcadeRepository extends JpaRepository<Arcade, Long> {
     Optional<Arcade> findArcadeByName(String name);
 }
